:wrench: Actually, improve this all around
